feat: added feature for handling of hpss commands
body: This checkin is to add a module for handling of hpss
commands starting with the htar -tvf command to inspect the
contents of tar files.  For this first checkin, all hpss
related functions are stored in one file, but as this grows,
the plan is to pull all the command-specific parsers and output
handlers into their own files.