feat: introduce external obs intervals on obs grouping plots
body: Sergey expressed a desire to compare the external observation
data sets availability to our observation archive stored on the
bdp bucket.  This pull request implements that capability.  Now,
'era5' and 'mera20cr' data set intervals will be overlayed onto
observation family subplots (orange line for era5 and blue line for
mera20cr).
In order to accommodate this change, a new section was added to
each observation family config yaml, the handling of the new config
settings was implemented, and the plotting routine was adjusted
accordingly.