feat: add sqlite db support for the new sinv command
body: also add foreign keys to several of the tables so that they
can be more easily related to other tables.  For instance, the
sinv command has a foreign key for the file id of the obs_inventory
table.  Likewise, now the obs_inventory table has a foreign key
referencing the cmd_results table.